requirements.txt
streamlit
pandas
openpyxl
emoji
pytz
gspread
google-auth-oauthlib